Moved theme.ts inside /src
