+++ packages/components/.storybook/preview.js
@@ -1,5 +1,5 @@
 import { ThemeProvider } from 'theme-ui'
-import { theme } from '../theme'
+import { theme } from '../src/theme'
 import '../src/assets/fonts/style.css'
 
 export const parameters = {
